avoid killing mypy
